read result return type
