+++ src/main/render/shared/identifiers.ts
@@ -37,6 +37,7 @@ export const COMMON_IDENTIFIERS = {
     encode: ts.createIdentifier('encode'),
     read: ts.createIdentifier('read'),
     write: ts.createIdentifier('write'),
+    failed: ts.createIdentifier('failed'),
     result: ts.createIdentifier('result'),
     val: ts.createIdentifier('val'),
     obj: ts.createIdentifier('obj'),
@@ -45,6 +46,7 @@ export const COMMON_IDENTIFIERS = {
     _requestId: ts.createIdentifier('_requestId'),
     err: ts.createIdentifier('err'),
     message: ts.createIdentifier('message'),
+    errMessage: ts.createIdentifier('errMessage'),
     args: ts.createIdentifier('args'),
     _args: ts.createIdentifier('_args'),
     fname: ts.createIdentifier('fname'),
@@ -84,6 +86,7 @@ export const COMMON_IDENTIFIERS = {
     Int64: ts.createIdentifier('thrift.Int64'),
     IHandler: ts.createIdentifier('IHandler'),
     ILocalHandler: ts.createIdentifier('ILocalHandler'),
+    IReadResult: ts.createIdentifier('IReadResult'),
     Node_Int64: ts.createIdentifier('Int64'),
     skip: ts.createIdentifier('skip'),
     writeMessageBegin: ts.createIdentifier('writeMessageBegin'),
